Version bump to 0.2.0

+++ lib/smartdc/api/keys.rb
@@ -2,16 +2,19 @@ module Smartdc
   module Api
     class Keys
       attr_reader :request
+      attr_reader :id
 
-      def initialize(request)
+      def initialize(request, id=nil)
         @request = request
+        @id = id
       end
 
       def create(params={})
         request.post('my/keys/', params)
       end
 
-      def read(id)  
+      def read
+        raise ArgumentError unless id
         request.get('my/keys/' + id.to_s)
       end
 
@@ -19,7 +22,8 @@ module Smartdc
         request.get('my/keys', params)
       end
 
-      def delete(id)  
+      def delete
+        raise ArgumentError unless id
         request.delete('my/keys/' + id.to_s)
         request.status == 204 ? true : false
       end
